fixed some bugs/errors, did more styling and improved the backend structure
