+++ src/models/user.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
-const saltRounds = 10;
+const hashPassword = require('../utils/hashPassword');
+const comparePassword = require('../utils/comparePassword');
 
 const userSchema = new mongoose.Schema({
 	username: {
@@ -50,42 +50,8 @@ const userSchema = new mongoose.Schema({
 	},
 });
 
-userSchema.pre('save', function (next) {
-	const user = this;
-
-	if (!user.isModified('password')) {
-		return next();
-	}
-
-	const hashPassword = (salt) => {
-		bcrypt.hash(user.password, salt, (err, hash) => {
-			if (err) {
-				console.log(err);
-			} else {
-				user.password = hash;
-				next();
-			}
-		});
-	};
-
-	bcrypt.genSalt(saltRounds, (err, salt) => {
-		if (err) {
-			console.log(err);
-		} else {
-			hashPassword(salt);
-		}
-	});
-});
-
-userSchema.methods.comparePassword = function (userPassword, callback) {
-	bcrypt.compare(userPassword, this.password, (err, matches) => {
-		if (err) {
-			console.log(err);
-		} else {
-			callback(undefined, matches);
-		}
-	});
-};
+userSchema.pre('save', hashPassword);
+userSchema.methods.comparePassword = comparePassword;
 
 const User = mongoose.model('users', userSchema);
