+++ src/static/styles/main-screen.css
@@ -631,7 +631,9 @@ nav {
   height: 100px;
   width: 100%;
   background-color: #F9F9F9;
-  padding: 30px; }
+  padding: 30px;
+  position: relative;
+  z-index: 2; }
   nav li {
     display: inline-block;
     margin: 0 10px; }
@@ -841,9 +843,60 @@ nav {
     font-family: 'Montserrat';
     font-size: 20px; }
 
+.m-background {
+  /* Permalink - use to edit and share this gradient: https://colorzilla.com/gradient-editor/#ffe7df+0,f8d4d9+100 */
+  background: #ffe7df;
+  /* Old browsers */
+  background: -moz-linear-gradient(top, #ffe7df 0%, #f8d4d9 100%);
+  /* FF3.6-15 */
+  background: -webkit-linear-gradient(top, #ffe7df 0%, #f8d4d9 100%);
+  /* Chrome10-25,Safari5.1-6 */
+  background: linear-gradient(to bottom, #ffe7df 0%, #f8d4d9 100%);
+  /* W3C, IE10+, FF16+, Chrome26+, Opera12+, Safari7+ */
+  filter: progid:DXImageTransform.Microsoft.gradient(startColorstr='#ffe7df', endColorstr='#f8d4d9', GradientType=0);
+  /* IE6-9 */
+  background-attachment: fixed;
+  z-index: -1; }
+  .m-background .blob-1 {
+    fill: #FFFBEA;
+    width: auto;
+    height: 800px;
+    opacity: .3;
+    position: absolute;
+    right: -100px;
+    top: -50px; }
+  .m-background .blob-2 {
+    fill: #FFFBEA;
+    width: 500px;
+    height: 800px;
+    opacity: .3;
+    position: absolute;
+    left: 30%;
+    top: 700px; }
+  .m-background .blob-3 {
+    fill: #FFFBEA;
+    width: 500px;
+    height: 800px;
+    opacity: .3;
+    position: absolute;
+    left: 70%;
+    top: 1200px; }
+  .m-background .logo-1 {
+    position: absolute;
+    right: -800px;
+    top: 500px;
+    z-index: 1; }
+  .m-background .logo-2 {
+    position: absolute;
+    left: -800px;
+    top: 1300px;
+    z-index: 1; }
+
 .m-footer {
   background-color: #100033;
-  min-height: 20vh; }
+  min-height: 20vh;
+  position: relative;
+  z-index: 2; }
   .m-footer p {
     margin: 0;
     padding-top: 50px;
@@ -857,14 +910,16 @@ nav {
 .p-register {
   width: 100%;
   padding: 50px 30px;
-  background-color: #FFE7DF;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
 
 .p-login {
   width: 100%;
   padding: 50px 30px;
-  background-color: #FFE7DF;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
   .p-login__form {
     animation: slideIn 600ms ease-in-out forwards; }
     .p-login__form button {
@@ -899,8 +954,9 @@ nav {
 .p-profile {
   width: 100%;
   padding: 50px 30px;
-  background-color: #FFE7DF;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
   .p-profile__avatar {
     position: absolute;
     width: 350px;
@@ -987,16 +1043,25 @@ nav {
         width: auto; } }
 
 .p-home {
-  background-color: #FFE7DF;
   padding-top: 100px;
   padding-bottom: 70px;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
   .p-home__banner {
     width: 100%;
     padding: 200px 0; }
     @media (max-width: 900px) {
       .p-home__banner {
         padding: 50px 0; } }
+    .p-home__banner img {
+      position: absolute;
+      transform: translateY(150px);
+      z-index: 1;
+      top: 0; }
+      @media (max-width: 900px) {
+        .p-home__banner img {
+          display: none; } }
     .p-home__banner__container {
       margin: 0 auto;
       max-width: 1000px;
@@ -1020,7 +1085,9 @@ nav {
         margin: 0 0 0 auto;
         font-family: 'Playfair Display';
         font-weight: bold;
-        font-size: 60px; }
+        font-size: 60px;
+        position: relative;
+        z-index: 2; }
         @media (max-width: 900px) {
           .p-home__banner__left h1 {
             width: 100%; } }
@@ -1066,20 +1133,23 @@ nav {
 .p-search {
   width: 100%;
   padding: 50px 30px;
-  background-color: #FFE7DF;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
 
 .p-matches {
   width: 100%;
   padding: 50px 30px;
-  background-color: #FFE7DF;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
 
 .p-other-profile {
   width: 100%;
   padding: 50px 30px;
-  background-color: #FFE7DF;
-  min-height: 80vh; }
+  min-height: 80vh;
+  position: relative;
+  z-index: 2; }
   .p-other-profile form {
     padding-left: 200px;
     width: 650px; }
@@ -1095,14 +1165,64 @@ nav {
       display: none; }
   .p-other-profile button {
     display: block; }
+  .p-other-profile a {
+    font-family: 'Montserrat';
+    font-weight: lighter;
+    margin-top: 15px;
+    width: 160px;
+    height: 50px;
+    border-radius: 6px;
+    color: #F9F9F9;
+    background-color: #100033;
+    border: 3px solid #100033;
+    transition: all 200ms ease-in-out;
+    text-decoration: none;
+    display: block;
+    padding-top: 12px;
+    text-align: center; }
+    .p-other-profile a:hover {
+      background-color: #F9F9F9;
+      color: #100033;
+      cursor: pointer; }
+  .p-other-profile .match-lightbox {
+    margin: 0 auto 70px;
+    width: 100%;
+    padding: 20px 70px;
+    color: #100033;
+    background-color: #F9F9F9;
+    font-family: 'Montserrat';
+    font-size: 18px;
+    font-weight: bold;
+    border-radius: 17.5px;
+    -webkit-box-shadow: 2px 2px 4px 1px rgba(0, 0, 0, 0.15);
+    -moz-box-shadow: 2px 2px 4px 1px rgba(0, 0, 0, 0.15);
+    box-shadow: 2px 2px 4px 1px rgba(0, 0, 0, 0.15);
+    width: 650px;
+    animation: slideIn 400ms ease-in-out forwards; }
+    .p-other-profile .match-lightbox h2 {
+      font-family: 'Playfair Display';
+      font-weight: bold;
+      font-size: 44px;
+      margin: 15px 0 0; }
+    @media (max-width: 1080px) {
+      .p-other-profile .match-lightbox {
+        width: 100%;
+        max-width: 770px; } }
+
+@keyframes slideIn {
+  0% {
+    transform: translateX(-100vw); }
+  100% {
+    transform: translateX(0vw); } }
 
 .p-404 {
-  background-color: #FFE7DF;
   padding-top: 50px;
   min-height: 80vh;
   text-align: center;
   font-family: 'Playfair Display';
-  font-weight: bold; }
+  font-weight: bold;
+  position: relative;
+  z-index: 2; }
   .p-404 h1 {
     font-size: 100px; }
   .p-404 h2 {
