+++ src/styles/modules/_m-background.scss
@@ -1,0 +1,60 @@
+.m-background {
+    /* Permalink - use to edit and share this gradient: https://colorzilla.com/gradient-editor/#ffe7df+0,f8d4d9+100 */
+    background: #ffe7df;
+    /* Old browsers */
+    background: -moz-linear-gradient(top, #ffe7df 0%, #f8d4d9 100%);
+    /* FF3.6-15 */
+    background: -webkit-linear-gradient(top, #ffe7df 0%, #f8d4d9 100%);
+    /* Chrome10-25,Safari5.1-6 */
+    background: linear-gradient(to bottom, #ffe7df 0%, #f8d4d9 100%);
+    /* W3C, IE10+, FF16+, Chrome26+, Opera12+, Safari7+ */
+    filter: progid:DXImageTransform.Microsoft.gradient(startColorstr='#ffe7df', endColorstr='#f8d4d9', GradientType=0);
+    /* IE6-9 */
+    // height: 100vh;
+    background-attachment: fixed;
+    z-index: -1;
+
+    .blob-1 {
+        fill: #FFFBEA;
+        width: auto;
+        height: 800px;
+        opacity: .3;
+        position: absolute;
+        right: -100px;
+        top: -50px;
+    }
+
+    .blob-2 {
+        fill: #FFFBEA;
+        width: 500px;
+        height: 800px;
+        opacity: .3;
+        position: absolute;
+        left: 30%;
+        top: 700px;
+    }
+
+    .blob-3 {
+        fill: #FFFBEA;
+        width: 500px;
+        height: 800px;
+        opacity: .3;
+        position: absolute;
+        left: 70%;
+        top: 1200px;
+    }
+
+    .logo-1 {
+        position: absolute;
+        right: -800px;
+        top: 500px;
+        z-index: 1;
+    }
+
+    .logo-2 {
+        position: absolute;
+        left: -800px;
+        top: 1300px;
+        z-index: 1;
+    }
+}
