+++ src/styles/modules/_m-nav.scss
@@ -3,6 +3,8 @@ nav {
     width: 100%;
     background-color: $white;
     padding: 30px;
+    position: relative;
+    z-index: 2;
 
     li {
         display: inline-block;
