+++ src/styles/pages/_p-home.scss
@@ -1,8 +1,9 @@
 .p-home {
-    background-color: $pink;
     padding-top: 100px;
     padding-bottom: 70px;
     min-height: 80vh;
+    position: relative;
+    z-index: 2;
 
     &__banner {
         width: 100%;
@@ -10,6 +11,17 @@
 
         @media (max-width: 900px) {
             padding: 50px 0;
+        }
+
+        img {
+            position: absolute;
+            transform: translateY(150px);
+            z-index: 1;
+            top: 0;
+
+            @media (max-width: 900px) {
+                display: none;
+            }
         }
 
         &__container {
@@ -41,6 +53,8 @@
                 font-family: 'Playfair Display';
                 font-weight: bold;
                 font-size: 60px;
+                position: relative;
+                z-index: 2;
 
                 @media (max-width: 900px) {
                     width: 100%;
