+++ src/styles/pages/_p-login.scss
@@ -1,8 +1,9 @@
 .p-login {
     width: 100%;
     padding: 50px 30px;
-    background-color: $pink;
     min-height: 80vh;
+    position: relative;
+    z-index: 2;
 
     &__form {
         animation: slideIn 600ms ease-in-out forwards;
