+++ src/styles/pages/_p-other-profile.scss
@@ -1,8 +1,9 @@
 .p-other-profile {
     width: 100%;
     padding: 50px 30px;
-    background-color: $pink;
     min-height: 80vh;
+    position: relative;
+    z-index: 2;
 
     form {
         padding-left: 200px;
@@ -21,6 +22,7 @@
     .tag-container-profile {
         .tag {
             padding-top: 8px;
+
             p {
                 display: none;
             }
@@ -29,5 +31,66 @@
 
     button {
         display: block;
+    }
+
+    a {
+        font-family: 'Montserrat';
+        font-weight: lighter;
+        margin-top: 15px;
+        width: 160px;
+        height: 50px;
+        border-radius: 6px;
+        color: $white;
+        background-color: $purple;
+        border: 3px solid $purple;
+        transition: all 200ms ease-in-out;
+        text-decoration: none;
+        display: block;
+        padding-top: 12px;
+        text-align: center;
+
+        &:hover {
+            background-color: $white;
+            color: $purple;
+            cursor: pointer;
+        }
+    }
+
+    .match-lightbox {
+        margin: 0 auto 70px;
+        width: 100%;
+        padding: 20px 70px;
+        color: $purple;
+        background-color: $white;
+        font-family: 'Montserrat';
+        font-size: 18px;
+        font-weight: bold;
+        border-radius: 17.5px;
+        -webkit-box-shadow: 2px 2px 4px 1px rgba(0, 0, 0, 0.15);
+        -moz-box-shadow: 2px 2px 4px 1px rgba(0, 0, 0, 0.15);
+        box-shadow: 2px 2px 4px 1px rgba(0, 0, 0, 0.15);
+        width: 650px;
+        animation: slideIn 400ms ease-in-out forwards;
+
+        h2 {
+            font-family: 'Playfair Display';
+            font-weight: bold;
+            font-size: 44px;
+            margin: 15px 0 0;
+        }
+
+        @media (max-width: 1080px) {
+            width: 100%;
+            max-width: 770px;
+        }
+    }
+
+    @keyframes slideIn {
+        0% {
+            transform: translateX(-100vw);
+        }
+        100% {
+            transform: translateX(0vw);
+        }
     }
 }
