+++ src/styles/pages/_p-profile.scss
@@ -1,8 +1,9 @@
 .p-profile {
     width: 100%;
     padding: 50px 30px;
-    background-color: $pink;
     min-height: 80vh;
+    position: relative;
+    z-index: 2;
 
     &__avatar {
         position: absolute;
