+++ src/styles/pages/_p-search.scss
@@ -1,6 +1,7 @@
 .p-search {
     width: 100%;
     padding: 50px 30px;
-    background-color: $pink;
     min-height: 80vh;
+    position: relative;
+    z-index: 2;
 }
