+++ src/utils/comparePassword.js
@@ -1,0 +1,13 @@
+const bcrypt = require('bcrypt');
+
+function comparePassword (userPassword, callback) {
+	bcrypt.compare(userPassword, this.password, (err, matches) => {
+		if (err) {
+			console.log(err);
+		} else {
+			callback(undefined, matches);
+		}
+	});
+}
+
+module.exports = comparePassword;
