+++ src/utils/userRedirectLogin.js
@@ -1,0 +1,9 @@
+const userRedirectLogin = (req, res, next) => {
+    if (!req.session.sessionID) {
+        res.redirect('/login');
+    } else {
+        next();
+    }
+};
+
+module.exports = userRedirectLogin;
