+++ src/utils/userRedirectProfile.js
@@ -1,0 +1,9 @@
+const userRedirectProfile = (req, res, next) => {
+    if (req.session.sessionID) {
+        res.redirect('/profile');
+    } else {
+        next();
+    }
+};
+
+module.exports = userRedirectProfile;
