+++ src/views/index.ejs
@@ -7,6 +7,7 @@
       <div class="p-home__banner__container">
         <div class="p-home__banner__left">
           <h1>Look at other people with the same interests!</h1>
+          <img src="/images/blob-main.svg" alt="">
         </div>
         <div class="p-home__banner__right">
           <h2>Check who you have already matched with!</h2>
@@ -22,7 +23,6 @@
     <% if (locals.user) { %>
     <%- include('includes/section_weather.ejs'); %>
     <% } %>
-
   </div>
 </main>
