+++ src/views/partials/shared/html-close.ejs
@@ -1,4 +1,4 @@
-    <!-- scripts -->
+    </div>
     <script src="/js/loginValidate.js"></script>
     <script src="/js/signupValidate.js"></script>
     <script src="/js/navMenu.js"></script>
