+++ src/views/partials/shared/html-open.ejs
@@ -23,3 +23,22 @@
 </head>
 
 <body>
+
+    <div class="m-background">
+        <svg class="blob-1">
+            <path
+                d="M241.460571,96.7880196 C264.102602,133.501764 275.69974,167.179034 276.251984,201.408389 C276.804229,235.637745 266.035458,270.143144 243.393427,302.716241 C220.751397,335.013294 186.236106,365.102002 138.466943,378.076032 C90.6977803,391.050062 29.9508683,386.909415 10.0700607,354.612361 C-9.81074685,322.315308 11.17455,261.585805 10.6223054,200.58026 C10.3461831,139.574714 -11.1913585,78.0170825 8.68944909,41.3033379 C28.5702567,4.58959335 90.1455357,-7.28026391 137.638576,6.5218957 C185.131616,20.6000985 218.81854,60.074275 241.460571,96.7880196 Z"
+                id="Path"></path>
+        </svg>
+        <svg class="blob-2">
+            <path
+                d="M371.804769,70.0560955 C423.352666,94.1617688 489.73449,130.876564 491.959579,169.816497 C493.81382,208.756431 431.511326,249.921504 379.592581,302.583129 C328.044684,355.61561 286.880536,420.886357 241.266209,425.707491 C195.651883,430.528626 144.845683,374.900149 98.4896601,321.867668 C51.7627894,269.206043 9.48609685,218.398701 1.32743689,160.174228 C-6.46037489,101.578899 20.6115423,35.5664398 66.9675648,11.4607665 C113.694435,-12.2740503 179.705411,5.89791883 231.253309,20.7321793 C283.172054,35.1955833 319.886024,46.3212787 371.804769,70.0560955 Z"
+                id="Path"></path>
+        </svg>
+        <svg class="blob-3">
+            <path
+                d="M241.460571,96.7880196 C264.102602,133.501764 275.69974,167.179034 276.251984,201.408389 C276.804229,235.637745 266.035458,270.143144 243.393427,302.716241 C220.751397,335.013294 186.236106,365.102002 138.466943,378.076032 C90.6977803,391.050062 29.9508683,386.909415 10.0700607,354.612361 C-9.81074685,322.315308 11.17455,261.585805 10.6223054,200.58026 C10.3461831,139.574714 -11.1913585,78.0170825 8.68944909,41.3033379 C28.5702567,4.58959335 90.1455357,-7.28026391 137.638576,6.5218957 C185.131616,20.6000985 218.81854,60.074275 241.460571,96.7880196 Z"
+                id="Path"></path>
+        </svg>
+        <img class="logo-1" src="/images/logo.png" alt="">
+        <img class="logo-2" src="/images/logo.png" alt="">
